systems/csv-to-json-converter/requirements.txt
# CSV-to-JSON Converter Dependencies

# Encoding detection
chardet>=5.2.0

# Date parsing for type inference
python-dateutil>=2.8.2
